# Stable Diffusion 3 support