… README.md
… configs/sd3-inference.yaml
… modules/models/sd3/sd3_impls.py
… modules/processing.py
… modules/prompt_parser.py
… modules/sd_hijack.py
… modules/sd_hijack_clip.py
… modules/sd_models.py
… modules/sd_models_config.py
… modules/sd_models_types.py
… modules/sd_samplers_common.py
… modules/sd_samplers_kdiffusion.py
… modules/sd_vae_approx.py
… modules/sd_vae_taesd.py
… modules/shared_options.py
… requirements.txt
… requirements_versions.txt