## Description

* initial SD3 support
* can load `sd3_medium.safetensors` from https://huggingface.co/stabilityai/stable-diffusion-3-medium
* will download CLIP models from huggingface into `models/CLIP` directory
* code for neural network, with minimal changes, taken from https://github.com/mcmonkey4eva/sd3-ref/
* added support for live previews with all three methods: cheap, approx-nn and TAESD
* uses a lot of VRAM
* I impersonated [mcmonkey4eva](https://github.com/mcmonkey4eva) in the commit a7116aa9a11c7d7d10df81aa52ad5ec33f4b6db6 so that the code is marked as contributed as him in git blame
* TODO:
- [x] add support for infinite length prompts
- [ ] add support for using optimized attention (currently uses SDP regardless of the setting, with some einsums inside) (do we even need that?)
- [x] add support for not loading T5
- [x] add support for counting tokens
- [ ] add support for DDIM and other timesteps profilers
- [ ] investigate why DPM++ 2M Karras isn't working well
- [ ] go through generation with profiler
- [x] support img2img
- [x] support medvram



## Screenshots/videos:

![firefox_D27dsqzVE7](https://github.com/AUTOMATIC1111/stable-diffusion-webui/assets/20920490/b8f2c239-0a64-4076-9e9c-528b5fb65525)