@protector131090 2024-06-16
AttributeError: 'NoneType' object has no attribute 'is_sdxl'