@protector131090 2024-06-16
> [protector131090](https://github.com/protector131090): Share the full stack trace from the console.

Traceback (most recent call last):
  File "C:\sd.webui\copy\venv\lib\site-packages\gradio\routes.py", line 488, in run_predict
    output = await app.get_blocks().process_api(
  File "C:\sd.webui\copy\venv\lib\site-packages\gradio\blocks.py", line 1429, in process_api
    inputs = self.preprocess_data(fn_index, inputs, state)
  File "C:\sd.webui\copy\venv\lib\site-packages\gradio\blocks.py", line 1222, in preprocess_data
    self.validate_inputs(fn_index, inputs)
  File "C:\sd.webui\copy\venv\lib\site-packages\gradio\blocks.py", line 1209, in validate_inputs
    raise ValueError(
ValueError: An event handler (modelmerger) didn't receive enough input values (needed: 16, got: 0).
Check if the event handler calls a Javascript function, and make sure its return value is correct.
Wanted inputs:
    [label, dropdown, dropdown, dropdown, radio, slider, checkbox, textbox, radio, radio, dropdown, textbox, checkbox, checkbox, checkbox, textbox]
Received inputs:
    []
Traceback (most recent call last):
  File "C:\sd.webui\copy\venv\lib\site-packages\gradio\routes.py", line 488, in run_predict
    output = await app.get_blocks().process_api(
  File "C:\sd.webui\copy\venv\lib\site-packages\gradio\blocks.py", line 1429, in process_api
    inputs = self.preprocess_data(fn_index, inputs, state)
  File "C:\sd.webui\copy\venv\lib\site-packages\gradio\blocks.py", line 1222, in preprocess_data
    self.validate_inputs(fn_index, inputs)
  File "C:\sd.webui\copy\venv\lib\site-packages\gradio\blocks.py", line 1209, in validate_inputs
    raise ValueError(
ValueError: An event handler (modelmerger) didn't receive enough input values (needed: 16, got: 0).
Check if the event handler calls a Javascript function, and make sure its return value is correct.
Wanted inputs:
    [label, dropdown, dropdown, dropdown, radio, slider, checkbox, textbox, radio, radio, dropdown, textbox, checkbox, checkbox, checkbox, textbox]
Received inputs:
    []