@protector131090 2024-06-16
> Well, that now does look like the right stack trace. Can you try running with environment variable PROTOCOL_BUFFERS_PYTHON_IMPLEMENTATION=python (i e add `set PROTOCOL_BUFFERS_PYTHON_IMPLEMENTATION=python` to webui-user.bat) to see if that changes anything?

venv "C:\sd.webui\copy\venv\Scripts\Python.exe"
Python 3.10.10 (tags/v3.10.10:aad5f6a, Feb  7 2023, 17:20:36) [MSC v.1929 64 bit (AMD64)]
Version: v1.9.4-173-g7ee2114c
Commit hash: 7ee2114cd9c401ceb390c141e604661055e0aaf4
[Auto-Photoshop-SD] Attempting auto-update...
[Auto-Photoshop-SD] switch branch to extension branch.
checkout_result: Your branch is up to date with 'origin/master'.

[Auto-Photoshop-SD] Current Branch.
branch_result: * master

[Auto-Photoshop-SD] Fetch upstream.
fetch_result:
[Auto-Photoshop-SD] Pull upstream.
pull_result: Already up to date.
Launching Web UI with arguments: --xformers --ckpt-dir S:\SD CHECKPOINTS --disable-safe-unpickle --api
python_server_full_path:  C:\sd.webui\copy\extensions\Auto-Photoshop-StableDiffusion-Plugin\server/python_server
[-] ADetailer initialized. version: 24.5.1, num models: 10
[AddNet] Updating model hashes...
0it [00:00, ?it/s]
[AddNet] Updating model hashes...
0it [00:00, ?it/s]
ControlNet preprocessor location: C:\sd.webui\webui\extensions\sd-webui-controlnet\annotator\downloads
2024-06-16 10:23:22,252 - ControlNet - INFO - ControlNet v1.1.449
Loading weights [912c9dc74f] from S:\SD CHECKPOINTS\realvisxlV40_v40Bakedvae.safetensors
Creating model from config: C:\sd.webui\copy\repositories\generative-models\configs\inference\sd_xl_base.yaml
Applying attention optimization: xformers... done.
Model loaded in 3.5s (load weights from disk: 0.3s, create model: 0.3s, apply weights to model: 2.5s, calculate empty prompt: 0.1s).
2024-06-16 10:23:26,357 - ControlNet - INFO - ControlNet UI callback registered.
C:\sd.webui\copy\extensions\sd-webui-additional-networks\scripts\metadata_editor.py:399: GradioDeprecationWarning: The `style` method is deprecated. Please set these arguments in the constructor instead.
  with gr.Row().style(equal_height=False):
C:\sd.webui\copy\extensions\sd-webui-additional-networks\scripts\metadata_editor.py:521: GradioDeprecationWarning: The `style` method is deprecated. Please set these arguments in the constructor instead.
  cover_image = gr.Image(
Running on local URL:  http://127.0.0.1:7860

To create a public link, set `share=True` in `launch()`.
Startup time: 25.6s (prepare environment: 11.0s, import torch: 3.0s, import gradio: 0.8s, setup paths: 0.8s, initialize shared: 0.2s, other imports: 0.4s, load scripts: 2.7s, create ui: 5.3s, gradio launch: 0.3s, add APIs: 0.9s).
100%|██████████████████████████████████████████████████████████████████████████████████| 20/20 [00:03<00:00,  6.59it/s]
Loading model comfy\sd3_medium.safetensors [cc236278d2] (2 out of 2)
Loading weights [cc236278d2] from C:\sd.webui\copy\models\Stable-diffusion\comfy\sd3_medium.safetensors
Creating model from config: C:\sd.webui\copy\configs\sd3-inference.yaml
mmdit initializing with: input_size=None, patch_size=2, in_channels=16, depth=24, mlp_ratio=4.0, learn_sigma=False, adm_in_channels=2048, context_embedder_config={'target': 'torch.nn.Linear', 'params': {'in_features': 4096, 'out_features': 1536}}, register_length=0, attn_mode='torch', rmsnorm=False, scale_mod_only=False, swiglu=False, out_channels=None, pos_embed_scaling_factor=None, pos_embed_offset=None, pos_embed_max_size=192, num_patches=36864, qk_norm=None, qkv_bias=True, dtype=torch.float16, device='cpu'
changing setting sd_model_checkpoint to comfy\sd3_medium.safetensors [cc236278d2]: OutOfMemoryError
Traceback (most recent call last):
  File "C:\sd.webui\copy\modules\options.py", line 165, in set
    option.onchange()
  File "C:\sd.webui\copy\modules\call_queue.py", line 14, in f
    res = func(*args, **kwargs)
  File "C:\sd.webui\copy\modules\initialize_util.py", line 181, in <lambda>
    shared.opts.onchange("sd_model_checkpoint", wrap_queued_call(lambda: sd_models.reload_model_weights()), call=False)
  File "C:\sd.webui\copy\modules\sd_models.py", line 950, in reload_model_weights
    sd_model = reuse_model_from_already_loaded(sd_model, checkpoint_info, timer)
  File "C:\sd.webui\copy\modules\sd_models.py", line 916, in reuse_model_from_already_loaded
    load_model(checkpoint_info)
  File "C:\sd.webui\copy\modules\sd_models.py", line 834, in load_model
    load_model_weights(sd_model, checkpoint_info, state_dict, timer)
  File "C:\sd.webui\copy\modules\sd_models.py", line 436, in load_model_weights
    model.load_state_dict(state_dict, strict=False)
  File "C:\sd.webui\copy\modules\sd_disable_initialization.py", line 223, in <lambda>
    module_load_state_dict = self.replace(torch.nn.Module, 'load_state_dict', lambda *args, **kwargs: load_state_dict(module_load_state_dict, *args, **kwargs))
  File "C:\sd.webui\copy\modules\sd_disable_initialization.py", line 221, in load_state_dict
    original(module, state_dict, strict=strict)
  File "C:\sd.webui\copy\venv\lib\site-packages\torch\nn\modules\module.py", line 2138, in load_state_dict
    load(self, state_dict)
  File "C:\sd.webui\copy\venv\lib\site-packages\torch\nn\modules\module.py", line 2126, in load
    load(child, child_state_dict, child_prefix)
  File "C:\sd.webui\copy\venv\lib\site-packages\torch\nn\modules\module.py", line 2126, in load
    load(child, child_state_dict, child_prefix)
  File "C:\sd.webui\copy\venv\lib\site-packages\torch\nn\modules\module.py", line 2126, in load
    load(child, child_state_dict, child_prefix)
  [Previous line repeated 7 more times]
  File "C:\sd.webui\copy\venv\lib\site-packages\torch\nn\modules\module.py", line 2120, in load
    module._load_from_state_dict(
  File "C:\sd.webui\copy\modules\sd_disable_initialization.py", line 225, in <lambda>
    linear_load_from_state_dict = self.replace(torch.nn.Linear, '_load_from_state_dict', lambda *args, **kwargs: load_from_state_dict(linear_load_from_state_dict, *args, **kwargs))
  File "C:\sd.webui\copy\modules\sd_disable_initialization.py", line 191, in load_from_state_dict
    module._parameters[name] = torch.nn.parameter.Parameter(torch.zeros_like(param, device=device, dtype=dtype), requires_grad=param.requires_grad)
  File "C:\sd.webui\copy\venv\lib\site-packages\torch\_meta_registrations.py", line 4507, in zeros_like
    res = aten.empty_like.default(
  File "C:\sd.webui\copy\venv\lib\site-packages\torch\_ops.py", line 448, in __call__
    return self._op(*args, **kwargs or {})
  File "C:\sd.webui\copy\venv\lib\site-packages\torch\_refs\__init__.py", line 4681, in empty_like
    return torch.empty_permuted(
torch.cuda.OutOfMemoryError: CUDA out of memory. Tried to allocate 160.00 MiB. GPU 0 has a total capacty of 23.99 GiB of which 0 bytes is free. Of the allocated memory 22.79 GiB is allocated by PyTorch, and 184.33 MiB is reserved by PyTorch but unallocated. If reserved but unallocated memory is large try setting max_split_size_mb to avoid fragmentation.  See documentation for Memory Management and PYTORCH_CUDA_ALLOC_CONF

Traceback (most recent call last):
  File "C:\sd.webui\copy\venv\lib\site-packages\gradio\routes.py", line 488, in run_predict
    output = await app.get_blocks().process_api(
  File "C:\sd.webui\copy\venv\lib\site-packages\gradio\blocks.py", line 1431, in process_api
    result = await self.call_function(
  File "C:\sd.webui\copy\venv\lib\site-packages\gradio\blocks.py", line 1103, in call_function
    prediction = await anyio.to_thread.run_sync(
  File "C:\sd.webui\copy\venv\lib\site-packages\anyio\to_thread.py", line 33, in run_sync
    return await get_asynclib().run_sync_in_worker_thread(
  File "C:\sd.webui\copy\venv\lib\site-packages\anyio\_backends\_asyncio.py", line 877, in run_sync_in_worker_thread
    return await future
  File "C:\sd.webui\copy\venv\lib\site-packages\anyio\_backends\_asyncio.py", line 807, in run
    result = context.run(func, *args)
  File "C:\sd.webui\copy\venv\lib\site-packages\gradio\utils.py", line 707, in wrapper
    response = f(*args, **kwargs)
  File "C:\sd.webui\copy\extensions\sd-webui-animatediff\scripts\animatediff_ui.py", line 273, in refresh_models
    new_model_list = self.get_model_list()
  File "C:\sd.webui\copy\extensions\sd-webui-animatediff\scripts\animatediff_ui.py", line 269, in get_model_list
    return [f for f in os.listdir(model_dir) if f != ".gitkeep" and not any(tag in f for tag in get_sd_rm_tag())]
  File "C:\sd.webui\copy\extensions\sd-webui-animatediff\scripts\animatediff_ui.py", line 269, in <listcomp>
    return [f for f in os.listdir(model_dir) if f != ".gitkeep" and not any(tag in f for tag in get_sd_rm_tag())]
  File "C:\sd.webui\copy\extensions\sd-webui-animatediff\scripts\animatediff_ui.py", line 261, in get_sd_rm_tag
    if shared.sd_model.is_sdxl:
AttributeError: 'NoneType' object has no attribute 'is_sdxl'
Traceback (most recent call last):
  File "C:\sd.webui\copy\venv\lib\site-packages\gradio\routes.py", line 488, in run_predict
    output = await app.get_blocks().process_api(
  File "C:\sd.webui\copy\venv\lib\site-packages\gradio\blocks.py", line 1431, in process_api
    result = await self.call_function(
  File "C:\sd.webui\copy\venv\lib\site-packages\gradio\blocks.py", line 1103, in call_function
    prediction = await anyio.to_thread.run_sync(
  File "C:\sd.webui\copy\venv\lib\site-packages\anyio\to_thread.py", line 33, in run_sync
    return await get_asynclib().run_sync_in_worker_thread(
  File "C:\sd.webui\copy\venv\lib\site-packages\anyio\_backends\_asyncio.py", line 877, in run_sync_in_worker_thread
    return await future
  File "C:\sd.webui\copy\venv\lib\site-packages\anyio\_backends\_asyncio.py", line 807, in run
    result = context.run(func, *args)
  File "C:\sd.webui\copy\venv\lib\site-packages\gradio\utils.py", line 707, in wrapper
    response = f(*args, **kwargs)
  File "C:\sd.webui\copy\extensions\sd-webui-animatediff\scripts\animatediff_ui.py", line 273, in refresh_models
    new_model_list = self.get_model_list()
  File "C:\sd.webui\copy\extensions\sd-webui-animatediff\scripts\animatediff_ui.py", line 269, in get_model_list
    return [f for f in os.listdir(model_dir) if f != ".gitkeep" and not any(tag in f for tag in get_sd_rm_tag())]
  File "C:\sd.webui\copy\extensions\sd-webui-animatediff\scripts\animatediff_ui.py", line 269, in <listcomp>
    return [f for f in os.listdir(model_dir) if f != ".gitkeep" and not any(tag in f for tag in get_sd_rm_tag())]
  File "C:\sd.webui\copy\extensions\sd-webui-animatediff\scripts\animatediff_ui.py", line 261, in get_sd_rm_tag
    if shared.sd_model.is_sdxl:
AttributeError: 'NoneType' object has no attribute 'is_sdxl'